update: remove unnecessary type casting for event flags in mouse and keyboard handling

+++ crates/bevy_cef_core/src/browser_process/browsers.rs
@@ -433,18 +433,19 @@ impl Browsers {
     }
 }
 
+#[allow(clippy::unnecessary_cast)]
 pub fn modifiers_from_mouse_buttons<'a>(buttons: impl IntoIterator<Item = &'a MouseButton>) -> u32 {
-    let mut modifiers = cef_event_flags_t::EVENTFLAG_NONE.0;
+    let mut modifiers = cef_event_flags_t::EVENTFLAG_NONE.0 as u32;
     for button in buttons {
         match button {
             MouseButton::Left => {
-                modifiers |= cef_event_flags_t::EVENTFLAG_LEFT_MOUSE_BUTTON.0
+                modifiers |= cef_event_flags_t::EVENTFLAG_LEFT_MOUSE_BUTTON.0 as u32
             }
             MouseButton::Right => {
-                modifiers |= cef_event_flags_t::EVENTFLAG_RIGHT_MOUSE_BUTTON.0
+                modifiers |= cef_event_flags_t::EVENTFLAG_RIGHT_MOUSE_BUTTON.0 as u32
             }
             MouseButton::Middle => {
-                modifiers |= cef_event_flags_t::EVENTFLAG_MIDDLE_MOUSE_BUTTON.0
+                modifiers |= cef_event_flags_t::EVENTFLAG_MIDDLE_MOUSE_BUTTON.0 as u32
             }
             _ => {}
         }
